response_schemas: move BackgroundJob schemas
