+++ cmk/gui/openapi/restful_objects/response_schemas.py
@@ -426,3 +426,34 @@ class AgentCollection(DomainObjectCollection):
         fields.Nested(AgentObject),
         description="A list of agent objects.",
     )
+
+
+class JobLogs(BaseSchema):
+    result = fields.List(
+        fields.String(),
+        description="The list of result related logs",
+    )
+    progress = fields.List(
+        fields.String(),
+        description="The list of progress related logs",
+    )
+
+
+class BackgroundJobStatus(BaseSchema):
+    active = fields.Boolean(
+        required=True,
+        description="This field indicates if the background job is active or not.",
+        example=True,
+    )
+    state = fields.String(
+        required=True,
+        description="This field indicates the current state of the background job.",
+        enum=["initialized", "running", "finished", "stopped", "exception"],
+        example="initialized",
+    )
+    logs = fields.Nested(
+        JobLogs,
+        required=True,
+        description="Logs related to the background job.",
+        example={"result": ["result1"], "progress": ["progress1"]},
+    )
